feat / Archive 사용자 공지사항 list, view

+++ src/main/java/kr/co/jhta/app/delideli/user/board/mapper/BoardMapper.java
@@ -46,7 +46,7 @@ public interface BoardMapper {
     void insertReplyComment(Comment comment);
 
     //내문의 글목록
-    List<Board> getMyAskList(Long userKey);
+    List<Board> getMyAskList(int userKey);
     //내문의 글작성
     void myAskWrite(Board board);
     //내문의 상세보기
